+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Changelog
 - Added: Django: Support for multiple databases (`Issue #9 <https://github.com/maxtepkeev/architect
   /issues/9>`__)
 - Changed: ``dsn`` partition option renamed to ``db`` to cover more use cases
-- Changed: ``DsnParseError`` renamed to ``OptionValueError`` to cover more use cases
+- Changed: ``DsnParseError`` exception renamed to ``OptionValueError`` to cover more use cases
 
 0.4.0 (2015-04-18)
 ++++++++++++++++++